pyproject.toml
# Poetry pyproject.toml: https://python-poetry.org/docs/pyproject/
[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "alpha_connector"
version = "0.1.1"
description = "Awesome `alpha_connector` is a Python cli/package created with https://github.com/TezRomacH/python-package-template"
readme = "README.md"
authors = ["Clovin4 <christian70401@gmail.com>"]
license = "MIT"
repository = "https://github.com/AlphaConnector/alpha_connector"
homepage = "https://github.com/AlphaConnector/alpha_connector"

# Keywords description https://python-poetry.org/docs/pyproject/#keywords
keywords = []  #! Update me

# Pypi classifiers: https://pypi.org/classifiers/
classifiers = [  #! Update me
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Operating System :: OS Independent",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
]

[tool.poetry.scripts]
# Entry points for the package https://python-poetry.org/docs/pyproject/#scripts
"alpha_connector" = "alpha_connector.__main__:app"

[tool.poetry.dependencies]
python = "^3.11"

typer = {extras = ["all"], version = "^0.9.0"}
rich = "^13.7.1"
python-dotenv = "^1.0.1"
pandas = "^2.2.1"
numpy = "^1.26.4"
xarray = "^2024.2.0"
pypdf = "^4.2.0"
matplotlib = "^3.8.4"
polars = "^0.20.25"
plotly = "^5.22.0"
nbformat = "^5.10.4"
pytest = "^8.2.2"

[tool.poetry.dev-dependencies]
bandit = "^1.7.1"
black = {version = "^24.3.0", allow-prereleases = true}
darglint = "^1.8.1"
isort = {extras = ["colors"], version = "^5.10.1"}
mypy = "^1.0"
mypy-extensions = "^1.0.0"
pre-commit = "^3.6.2"
pydocstyle = "^6.1.1"
pylint = "^3.1.0"
pytest = "^8.1.1"
pyupgrade = "^3.15.1"
safety = "^3.0.1"
coverage = "^7.4.4"
coverage-badge = "^1.1.0"
pytest-html = "^4.1.1"
pytest-cov = "^4.1.0"
click = "^8.1.7"

[tool.poetry.group.dev.dependencies]
requests-mock = "^1.11.0"
ipykernel = "^6.29.4"

[tool.black]
# https://github.com/psf/black
target-version = ["py39"]
line-length = 88
color = true

exclude = '''
/(
    \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | env
    | venv
)/
'''

[tool.isort]
# https://github.com/timothycrosley/isort/
py_version = 39
line_length = 88

known_typing = ["typing", "types", "typing_extensions", "mypy", "mypy_extensions"]
sections = ["FUTURE", "TYPING", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
include_trailing_comma = true
profile = "black"
multi_line_output = 3
indent = 4
color_output = true

[tool.mypy]
# https://mypy.readthedocs.io/en/latest/config_file.html#using-a-pyproject-toml-file
python_version = "3.11"
pretty = true
show_traceback = true
color_output = true

allow_redefinition = false
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
ignore_missing_imports = true
implicit_reexport = false
no_implicit_optional = true
show_column_numbers = true
show_error_codes = true
show_error_context = true
strict_equality = true
strict_optional = true
warn_no_return = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
warn_unused_configs = true
warn_unused_ignores = true


[tool.pytest.ini_options]
# https://docs.pytest.org/en/6.2.x/customize.html#pyproject-toml
# Directories that are not visited by pytest collector:
norecursedirs =["hooks", "*.egg", ".eggs", "dist", "build", "docs", ".tox", ".git", "__pycache__"]
doctest_optionflags = ["NUMBER", "NORMALIZE_WHITESPACE", "IGNORE_EXCEPTION_DETAIL"]

# Extra options:
addopts = [
  "--strict-markers",
  "--tb=short",
  "--doctest-modules",
  "--doctest-continue-on-failure",
]

[tool.coverage.run]
source = ["tests"]


[tool.coverage.paths]
source = ["alpha_connector"]

[tool.coverage.report]
fail_under = 50
show_missing = true
